Refactor models and fix N N+1 bug
