+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/KeyFrameType.cs
@@ -6,6 +6,7 @@
         //No longer used as it overlapped with UI_ELEMENT/GAME_ELEMENT
         SCENE,
         GAME_ELEMENT,
+        //No longer used as it overlapped with GAME_ELEMENT when we went to bot segments and removed roll up of objects
         GAME_ELEMENT_RENDERER_COUNT,
         UI_ELEMENT,
         //TODO: Implement me for bot segments
