+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/BaseFrameStateData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 87e5da71de784ba2aff5fac2306fb01b
+timeCreated: 1717696754
