+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/BaseReplayObjectState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2d614dd0869b46828b84096cb29af0b2
+timeCreated: 1717696794
