+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/BehaviourState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f983d94c51ec43169ac181d9a7ab4112
+timeCreated: 1717696866
