+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/Collider2DRecordState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 26d70e046bf148c69dba1cfc013c1ef1
+timeCreated: 1717696910
