+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/ColliderRecordState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3b616b0bb31441cf92afe313b10494eb
+timeCreated: 1717696885
