+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/ColliderReplayState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 30fc8fc8778b4e818e82e68d5d8ce2c7
+timeCreated: 1717696927
