+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/EngineStatsData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 51f232688fa442ddb31bbb8456a00603
+timeCreated: 1717696726
