+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/RecordedGameObjectState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2489802eaee24c59b334365792951e85
+timeCreated: 1717696846
