+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/RecordingFrameStateData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fb907cac0a394a04b05b678253f23c75
+timeCreated: 1717696677
