+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/ReplayFrameStateData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 999f9db178934a948a053c9537e45939
+timeCreated: 1717696774
