+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/ReplayGameObjectState.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Diagnostics.CodeAnalysis;
+
+namespace RegressionGames.StateRecorder.Models
+{
+    [Serializable]
+    [SuppressMessage("ReSharper", "InconsistentNaming")]
+    // replay doesn't need to deserialize everything we record
+    public class ReplayGameObjectState : BaseReplayObjectState
+    {
+    }
+}
