+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/ReplayGameObjectState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4bb32490104641d2b26c9e1f61cb881d
+timeCreated: 1717696823
