+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/Rigidbody2DRecordState.cs
@@ -1,0 +1,43 @@
+﻿using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Text;
+using RegressionGames.StateRecorder.JsonConverters;
+using UnityEngine;
+
+namespace RegressionGames.StateRecorder.Models
+{
+    [Serializable]
+    [SuppressMessage("ReSharper", "InconsistentNaming")]
+    public class Rigidbody2DRecordState: RigidbodyRecordState
+    {
+
+        // keep a ref to this instead of updating fields every tick
+        public new Rigidbody2D rigidbody;
+
+        public override void WriteToStringBuilder(StringBuilder stringBuilder)
+        {
+            stringBuilder.Append("{\"path\":");
+            StringJsonConverter.WriteToStringBuilder(stringBuilder, path);
+            stringBuilder.Append(",\"normalizedPath\":");
+            StringJsonConverter.WriteToStringBuilder(stringBuilder, normalizedPath);
+            stringBuilder.Append(",\"is2D\":true");
+            stringBuilder.Append(",\"position\":");
+            VectorJsonConverter.WriteToStringBuilderVector3(stringBuilder, rigidbody.position);
+            stringBuilder.Append(",\"rotation\":");
+            QuaternionJsonConverter.WriteToStringBuilder(stringBuilder, Quaternion.Euler(0, 0, rigidbody.rotation));
+            stringBuilder.Append(",\"velocity\":");
+            VectorJsonConverter.WriteToStringBuilderVector3(stringBuilder, rigidbody.velocity);
+            stringBuilder.Append(",\"mass\":");
+            FloatJsonConverter.WriteToStringBuilder(stringBuilder, rigidbody.mass);
+            stringBuilder.Append(",\"drag\":");
+            FloatJsonConverter.WriteToStringBuilder(stringBuilder, rigidbody.drag);
+            stringBuilder.Append(",\"angularDrag\":");
+            FloatJsonConverter.WriteToStringBuilder(stringBuilder, rigidbody.angularDrag);
+            stringBuilder.Append(",\"gravityScale\":");
+            FloatJsonConverter.WriteToStringBuilder(stringBuilder, rigidbody.gravityScale);
+            stringBuilder.Append(",\"isKinematic\":");
+            stringBuilder.Append((rigidbody.isKinematic ? "true" : "false"));
+            stringBuilder.Append("}");
+        }
+    }
+}
