+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/Rigidbody2DRecordState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0525dc0d4d0443329262b29aaea7cead
+timeCreated: 1717696966
