+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/RigidbodyRecordState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b603b2e31ee24964b4f6c19196c4d9fb
+timeCreated: 1717696946
