+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/RigidbodyReplayState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 776d9e513e7a4fe9bfe67cb7a66cdabf
+timeCreated: 1717696990
