--- src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/RecordingAndReplayContainerTypes.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 641a0f0a0dc64ce4bac4807174ef3ed1
-timeCreated: 1709049805
