+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/ScreenRecorder.cs
@@ -324,7 +324,7 @@ namespace RegressionGames.StateRecorder
         // cache this to avoid re-alloc on every frame
         private readonly List<KeyFrameType> _keyFrameTypeList = new(10);
 
-        private void GetKeyFrameType(bool firstFrame, bool hasUIDelta, bool hasGameObjectDelta, bool hasRendererDelta, string pixelHash)
+        private void GetKeyFrameType(bool firstFrame, bool hasUIDelta, bool hasGameObjectDelta, string pixelHash)
         {
             _keyFrameTypeList.Clear();
             if (firstFrame)
@@ -346,11 +346,6 @@ namespace RegressionGames.StateRecorder
             if (hasGameObjectDelta)
             {
                 _keyFrameTypeList.Add(KeyFrameType.GAME_ELEMENT);
-            }
-
-            if (hasRendererDelta)
-            {
-                _keyFrameTypeList.Add(KeyFrameType.GAME_ELEMENT_RENDERER_COUNT);
             }
         }
 
@@ -442,14 +437,14 @@ namespace RegressionGames.StateRecorder
                 }
 
                 //Compute the delta values we need to record / evaluate to know if we need to record a key frame
-                var uiDeltas = InGameObjectFinder.GetInstance().ComputeNormalizedPathBasedDeltaCounts(uiTransforms.Item1, uiTransforms.Item2, out var hasUIDelta, out var _);
-                var gameObjectDeltas = InGameObjectFinder.GetInstance().ComputeNormalizedPathBasedDeltaCounts(gameObjectTransforms.Item1, gameObjectTransforms.Item2, out var hasGameObjectDelta, out var hasGameObjectRendererDelta);
+                var uiDeltas = InGameObjectFinder.GetInstance().ComputeNormalizedPathBasedDeltaCounts(uiTransforms.Item1, uiTransforms.Item2, out var hasUIDelta);
+                var gameObjectDeltas = InGameObjectFinder.GetInstance().ComputeNormalizedPathBasedDeltaCounts(gameObjectTransforms.Item1, gameObjectTransforms.Item2, out var hasGameObjectDelta);
 
                 var gameFacePixelHashObserver = GameFacePixelHashObserver.GetInstance();
                 var pixelHash = gameFacePixelHashObserver != null ? gameFacePixelHashObserver.GetPixelHash(true) : null;
 
                 // tell if the new frame is a key frame or the first frame (always a key frame)
-                GetKeyFrameType(_tickNumber ==0, hasUIDelta, hasGameObjectDelta, hasGameObjectRendererDelta, pixelHash);
+                GetKeyFrameType(_tickNumber ==0, hasUIDelta, hasGameObjectDelta, pixelHash);
 
                 BotSegment botSegment = null;
                 // estimating the time in int milliseconds .. won't exactly match target FPS.. but will be close
@@ -509,9 +504,7 @@ namespace RegressionGames.StateRecorder
                                     count =  a.count,
                                     addedCount = a.addedCount,
                                     removedCount = a.removedCount,
-                                    countRule = a.higherLowerCountTracker ==0 ? (a.count==0 ? CountRule.Zero : CountRule.NonZero) : (a.higherLowerCountTracker > 0 ? CountRule.GreaterThanEqual : CountRule.LessThanEqual),
-                                    rendererCount = a.rendererCount,
-                                    rendererCountRule = a.higherLowerRendererCountTracker ==0 ? (a.rendererCount==0 ? CountRule.Zero : CountRule.NonZero) : (a.higherLowerRendererCountTracker > 0 ? CountRule.GreaterThanEqual : CountRule.LessThanEqual),
+                                    countRule = a.higherLowerCountTracker ==0 ? (a.count==0 ? CountRule.Zero : CountRule.NonZero) : (a.higherLowerCountTracker > 0 ? CountRule.GreaterThanEqual : CountRule.LessThanEqual)
                                 }
                             })
                             .ToArray();
